Unmount the container and staging paths of the volumes during deletion
Fixes #90
